requirements.txt
streamlit>=1.36
Pillow>=10.0.0
numpy>=1.26
opencv-python-headless>=4.9
imageio[ffmpeg]>=2.34
edge-tts>=6.1.12
python-dotenv>=1.0
